requirements.txt
autopep8==1.5.5
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
idna==2.10
pycodestyle==2.6.0
pyzmq==22.0.3
requests==2.25.1
soupsieve==2.2
toml==0.10.2
urllib3==1.26.3
Wikipedia-API==0.5.4
